requirements.txt
streamlit>=1.20
openai>=0.27.0
python-dotenv>=1.0.0
requests>=2.28.0
